Card table as DCQ
* gc thread cpu time tracking with gc+cpu=debug logging

* use correct young gen RS length prediction for base time calculation after finalizing young gen

* proper accounting of sweep rt continuation to base time
* merge to actual remset size calculation fixes

* fix compilation on aarch64, Windows

* fix compilation on aarch64

* remove trailing spaces from files

* initial riscv implementation
* cleanup s390/ppc
* fix missing ResizeTLABs event in jfr parallel phases test

* ppc barrier attempt (not even built)

* removed necessary C1 slow path stub generation :(

* fix riscv compilation

* fix RISCV barrier, passes javac HelloWorld and its execution

* re-add too-much-deleted stuff
* consider yield time in dirtying rate calculations
* cleanup

* clean up in refinement heuristics

* remove unused g1_young_card_val

* refactoring, renaming

* when calculating total merged cards from RS, compensate for the merge cache
* cleanup, refactoring, renaming

* refactoring of refinement/redirtying stats recording
* disable some expensive logging

* fix merge card cache compensation

* improve documentation about CardValue's LSB discriminating between clean/non-clean
* additional comments to assembly code

* fix too early clearing of refinement statistics after regular refinement completion where cards_to_cset would always be zero
* improved sizing of work for refinement table merge phase
* refactoring, removing fixmes

* aarch64 improved array post barrier

* fix aarch64 array post barrier assembler version

* fix testgclogmessages.java test after recent changes

* arm32 barrier

* currently yield duration only applies to sweeping

* add missing files after rebase

* regularize new_val_maybe_null * ppc build fixes

* more ppc build fixes after bad merge

* too many fixes :(

* cleanup

* fix check for enough space to evacuate

* remove code to take expected old gen surviving words into account when determining eden length

* remove some debug code

* track safepoints in recent refinement epoch to calculate card dirtying time

* remove card_table1 member from g1barrierset

* refactoring, cleanup

* fix issues with tracking gc pauses for card dirtying

* epoch timing fixes; little cleanup

* more time accounting fixes

* some refactoring

* cleanup

* remove more debug logs
* remove parts of already pushed stuff before merge

* improve cpu time output

* synchronize accesses for prediction relevant members between refinement and young gen revise thread

* remove UseNewCode in barrier code
* some assert to check that dirtying cards is done at the right time

* comment why the lock when updating redirtying information

* remove dead code

* some cleanup in code generation

* initial version

* add store_addr == new_val check to all platforms

* remove unnecessary stuff

* fix ppc barrier code (from M. Doerr)
  * too strong different register assertion due to ppc optimization
* fix passing of new_val_may_be_null for c1 barriers

* factor out x.a = x assignments for the C1 compiler.

* remove FIXMEs

* fix s390 barrier code (from A. Kumar)

* removed empty JMVCI write_barrier_post stub because JVMCI users need more changes than that anyway
* added card table base offset constant for use with JVMCI

* add clean_card_val() for JVMCI

* fix node costs for g1 post barrier

* remove trailing whitespace in files

* update post barrier cost estimate

* copyright updates
* cleanup of G1CardTableClaimTable
* worker threads elapsed CPU time refactoring
* undid _maybe_null from _may_be_null in code generator
* minor dead code removal

* some cleanup

* minor refactorings

* update cpu time gathering to proposed upstream

* remove cpu time logging code - can be retrieved from performance counters as well

* renamings of "primary control thread" to "refinement control thread" everywhere
* simplify refinement control thread main loop

* add refinement control thread perf counters

+++ src/hotspot/share/gc/g1/g1SurvRateGroup.cpp
@@ -119,7 +119,7 @@ void G1SurvRateGroup::fill_in_last_surv_rates() {
 void G1SurvRateGroup::finalize_predictions(const G1Predictions& predictor) {
   double accum = 0.0;
   double pred = 0.0;
-  for (size_t i = 0; i < _stats_arrays_length; ++i) {
+  for (uint i = 0; i < _stats_arrays_length; ++i) {
     pred = predictor.predict_in_unit_interval(_surv_rate_predictors[i]);
     accum += pred;
     _accum_surv_rate_pred[i] = accum;
